Big commit with working non-diggle leaveoneout
In the end, the best solution was to use a c function for the
excess, and use the (spatstat) density.ppp to compute the double
sum. Lesson: two single sums is a lot faster than one double sum.
I think it may be the number of exps that really makes the difference.

Note: pcfinhom and Kglobal are implemented, the cross versions are not.